Removed linalg.py
